complete: Alternate completion using readdir()
In an embedded system, the technique of spawning an "echo" to obtain
matching filenames may not apply.  An alternative completion scheme
uses opendir() / readdir() to find matches.  There are differences
in behaviour: (1) no generic wildcard expansion; (2) matching names
are not returned in sorted order; (3) filenames containing spaces
and shell metacharacters are handled correctly.